+++ src/SMTLIB/Backends.hs
@@ -1,5 +1,4 @@
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE TupleSections #-}
 
 module SMTLIB.Backends
   ( Backend (..),
@@ -14,7 +13,7 @@ where
 import Data.ByteString.Builder (Builder)
 import qualified Data.ByteString.Lazy.Char8 as LBS
 import Data.Char (isSpace)
-import Data.IORef (IORef, atomicModifyIORef, newIORef)
+import Data.IORef (IORef, modifyIORef, newIORef, readIORef, writeIORef)
 import Data.List (intersperse)
 import Prelude hiding (log)
 
@@ -22,6 +21,8 @@ import Prelude hiding (log)
 -- processes them and outputs the solver's response.
 newtype Backend = Backend
   { -- | Send a command to the backend.
+    -- While the implementation depends on the backend, this function is usually
+    -- *not* thread-safe.
     send :: Builder -> IO LBS.ByteString
   }
 
@@ -33,14 +34,18 @@ data QueuingFlag = Queuing | NoQueuing
 -- | Push a command on the solver's queue of commands to evaluate.
 -- The command must not produce any output when evaluated, unless it is the last
 -- command added before the queue is flushed.
+-- For a fixed queue, this function is *not* thread-safe.
 putQueue :: Queue -> Builder -> IO ()
-putQueue q cmd = atomicModifyIORef q $ \cmds ->
-  (cmds <> cmd, ())
+putQueue q cmd = modifyIORef q (<> cmd)
 
 -- | Empty the queue of commands to evaluate and return its content as a bytestring
 -- builder.
+-- For a fixed queue, this function is *not* thread-safe.
 flushQueue :: Queue -> IO Builder
-flushQueue q = atomicModifyIORef q (mempty,)
+flushQueue q = do
+  cmds <- readIORef q
+  writeIORef q mempty
+  return cmds
 
 -- | A solver is essentially a wrapper around a solver backend. It also comes with
 -- a function for logging the solver's activity, and an optional queue of commands
@@ -103,6 +108,7 @@ initSolver queuing solverBackend = do
 -- | Have the solver evaluate a SMT-LIB command.
 -- This forces the queued commands to be evaluated as well, but their results are
 -- *not* checked for correctness.
+-- For a fixed backend, this function is *not* thread-safe.
 command :: Solver -> Builder -> IO LBS.ByteString
 command solver cmd = do
   send (backend solver)
@@ -115,6 +121,7 @@ command solver cmd = do
 -- In lazy mode, (unless the queue is flushed and evaluated
 -- right after) the command must not produce any output when evaluated, and
 -- its output is thus in particular not checked for correctness.
+-- For a fixed backend, this function is *not* thread-safe.
 command_ :: Solver -> Builder -> IO ()
 command_ solver cmd =
   case queue solver of